Modified check pixels
Now it checks the input answer. You can now select the value and fastly correct it while looking at the mistake
with an image viewer.
It also checks the input number is inside the look up table selected to not fail during conversion.
Organizing label.py so that the 3 sc labels added are together